Initial commit
Say hello to Quartz, a modern QML library.

This also adds a small demo to show how to link against quartz.
According to the docs [0] we are also required to do something like:

```
#include <QtQml/QQmlExtensionPlugin>
Q_IMPORT_QML_PLUGIN(QuartzPlugin)
```

But in my testing this was not required. Just leaving this info here in
case it is helpful later on.

We should also look into how we properly place this qml module in the
resources [1]. We have to test if naming collisions are indeed possible,
otherwise the default resource path would work just fine.

[0] https://doc.qt.io/qt-6/qtqml-writing-a-module.html
[1] https://doc.qt.io/qt-6/qtqml-writing-a-module.html#eliminating-run-time-file-system-access