Add singleton macros
The QML_CPP_SINGLETON macro is still a hack, because of Qt's ridiculous
default behaviour to always prefer the default constructor even if the
create() function with the correct signature exists.

In the future we might want to allow overriding the empty default
constructor, by allowing a second parameter for the macro that inserts
its argument inside the default constructor definition.