Workaround qmlimportscanner being slow
As stated in 986ab495575ea1065985d01f6cdf731b28f29143, without any
workarounds qmlimportscanner will scan all directories recursively [0]
from beginning from the root of the repository, which is obviously
terribly slow if it contains a directory with thousands of icon files
like material-design-icons.

We can speed up things a little bit by calling the icon subdirectory
"Debug-iphoneos" [1], but it does not help a whole lot, because
QDirIterator still iterates over all files and directories, just the
inner-most block is cut short.

Well this is all quite the mess and unfortunately there is no option in
qmlimportscanner to ignore directories. But that would not be a good
solution either if it existed, because we would require all users of
this library to add an ignore directory in their own (possible)
qmlimportscanner calls. And there are a lot of implicit qmlimportscanner
calls in Qt cmake land.

There has been some performance improvements in qmlimportscanner in the
past [2], but they do not really help in our case.

We might want to fix this upstream in a better way, but for now we
settle with this workaround:
- After FetchContent has finished downloading, copy all (and only) the
  icons that we need
- Copy them to a directory with the name "Debug-iphoneos" in it, so that
  we get qmlimportscanner to at least skip them (this helps if the user
  bundles many icons)
- Delete the whole source directory downloaded by FetchContent
  afterwards, we do not need it anymore as we copied all relevant files

This workaround makes up for quite the nice build speedup again.

Before this changing a single line in Main.qml would cause the iterative
rebuild to take ~10 seconds, now it is 1-2 seconds.

[0] https://code.qt.io/cgit/qt/qtdeclarative.git/tree/tools/qmlimportscanner/main.cpp?h=v6.6.1#n727
[1] https://code.qt.io/cgit/qt/qtdeclarative.git/tree/tools/qmlimportscanner/main.cpp?h=v6.6.1#n742
[2] https://codereview.qt-project.org/c/qt/qtdeclarative/+/412828