Switch from material icons to material symbols
This is the more modern Material icon variant that Google is developing.
Unfortunately the official Google repo is not really usable. Even with a
shallow clone the repo has size reaching a gigabyte, which is
unacceptable as we need to download it (just to throw most of the icons
away afterwards).

Instead we now use a fork by marella, which is much smaller (download
size 41MB). This is not an official Google endorsed repo, hence we
secure the download with a SHA256 match and try to audit the repo again
every time we update it in the future.

Note that the reconfigure time of the demo is a bit slow, but this is
only because we bundle ~6000 icon resources, which Qt's internal cmake
code has to process during configure time. During compile time (and
runtime) there is not really a large impact from bundling that many
icons, and the reconfigure problem does not exist if an application
bundles a reasonable amount of icons.

The update to material symbols also gets rid of the icon category
concept from material icons: Everything lives in the same category now.

This patch also moves all icons to the "/svg" prefix in the resource
system. It is not much more to type (and certainly not since we remove
icon categories with this patch) but it is a tad more organized.