Bundle icons as font instead of as multiple SVGs
Previously I always had doubt if it makes sense to bundle icons as a
font, but after this rewrite I am confident that this approach is much
more sane.

Other modern technologies have been using icons like this for quite a
while and it makes integration just so much easier.
Also with Qt 6.7 having support for variable fonts [0] we can even use
some extra features such as animating the icons with the variable font
axes (e.g. weight).

With this we can also get rid of all the extra options related to how
many icons we want to bundle, the font is just 3MB large after all
(maybe in the future we can still add an option to allow not bundling
this font in case it is really not needed).
And thus we stop benchmarking cmake and qt_add_resources() by no longer
adding 6000 icons to the resource system, which allows us to get rid of
some cursed cmake workarounds to keep configure times low.

Note that the icons page now has a small regression where icons are not
sorted anymore. This is obviously because we cache the codepoints in a
hashmap, but I do not really care about this as noone probably wants to
list all icons in production: Fast lookup of a specific icon is much
more important.

[0] https://bugreports.qt.io/browse/QTBUG-108624