Add option to prevent bundling icons
People can use this option if they do not need icons, but it does not
save much space (about 3MB).

Using this option will also at the moment generate a warning at runtime:
QML Icons: Cannot load font: "qrc:/MaterialSymbols.woff2"

We could fix this warning in the future with a binding to an ifdefed
property in C++, but it is also safe to ignore the warning.