Attempt to stretch IconButton text to width
We should probably also remove the width stretching from the icon, so
that the text can get most of the width. But this will require some
fine-tuning to make sure that the icon-only variant still keeps the icon
centered, and that the icon is not completely left-anchored (need some
margin) for the icon+text variant.