Enable variableAxes feature
This requires Qt 6.7.

Right now this is still broken due to a component versioning bug [0],
but we enable it anyway now, as the upstream bug will likely be fixed
soon.

We should also consider if we want to enable the new
"Text.CurveRendering" for this, also refer to the related blog post [1].

[0] https://bugreports.qt.io/browse/QTBUG-124359
[1] https://www.qt.io/blog/text-improvements-in-qt-6.7