nix: Courtesy export src
This makes it easier to point to the src for FetchContent in Nix.
Technically the src is already part of outputs, but the attribute path
would be quite long with "quartz.packages.${system}.default.src".
With this short courtesy export it becomes just "quartz.src" and we skip
the Nix flakes $system bullshit as well.