nix: Export icon derivations
This makes it much easier to bundle FetchContent icons in flakes.
The cmakeFlags utility also automatically adds the right cmake flags.
The implementation was a bit cursed, mostly because of our overly eager
Regex matching, but also because downloading single files in Nix is hard
for no reason [0]. Normally you would go for pkgs.fetchurl, but since we
need to reuse the hashes from cmake, and because pkgs.fetchurl does some
transformations to the hashes, we need to use builtins.fetchurl so that
we can work with the same SHA256 hash, that sha256sum would give us.
Thus, further derivation wrappers are needed.

[0] https://github.com/NixOS/nixpkgs/issues/93520