Adopt recommended QML directory layout
The official documentation recommends the QML directory layout to follow
the name of the QML module (replacing "." with "/") [0], there is no
room for allowing all QML files to be placed cleanly in a separate qml
directory.
I would much prefer it that way, but unfortunately recently Qt
introduced some warnings, e.g. related to QTP0004 [1], if we don't
follow this hideous practice.

For the moment, I can't be bothered to fix QTP0004 for the example
projects, as it's just more practical to do everything in a single
CMakeLists.txt in that case, plus it is ridiculous to create an extra
QML module plus an entire directory layout for what is essentially just
a Main.qml.

[0] https://doc.qt.io/qt-6/qt-add-qml-module.html#separate-backing-and-plugin-targets
[1] https://doc.qt.io/qt-6/qt-cmake-policy-qtp0004.html