Prefer ttf icon font
This improves compatibility as woff2 is not supported on all platforms,
for example not on freetype on Android.
If Android ever gets support for woff2, we should revert this change as
woff2 is about three times smaller than ttf.

In order to not increase the complexity regarding icon download
checksums, we have to change the default for all platforms, which is
rather unfortunate.